+++ tests/Unit/StarterBase/RegisterMiscHooksTest.php
@@ -90,7 +90,7 @@ class RegisterMiscHooksTest extends StarterBaseTestCase {
 		$this->invokeRegisterMiscHooks( $this->bareInstance() );
 
 		$this->assertSame(
-			[ 'run_wptexturize', 'wpcf7_autop_or_not', 'wpml_tm_translation_job_data', 'option_icl_sitepress_settings' ],
+			[ 'run_wptexturize', 'wpcf7_autop_or_not', 'wpml_tm_translation_job_data', 'option_icl_st_settings', 'default_option_icl_st_settings', 'pre_update_option_icl_st_settings' ],
 			$filters
 		);
 	}
@@ -103,7 +103,7 @@ class RegisterMiscHooksTest extends StarterBaseTestCase {
 
 		$this->invokeRegisterMiscHooks( $this->bareInstance() );
 
-		$this->assertContains( 'option_icl_sitepress_settings', $filters );
+		$this->assertContains( 'option_icl_st_settings', $filters );
 	}
 
 	public function test_does_not_register_wpml_theme_domain_authoritative_filter_when_flag_disabled(): void {
@@ -118,6 +118,6 @@ class RegisterMiscHooksTest extends StarterBaseTestCase {
 
 		$this->invokeRegisterMiscHooks( $instance );
 
-		$this->assertNotContains( 'option_icl_sitepress_settings', $filters );
+		$this->assertNotContains( 'option_icl_st_settings', $filters );
 	}
 }